+++ RacingLeagueHub.Application/Identity/Authentication/TwoFactorAuthentication/ITwoFactorService.cs
@@ -4,7 +4,7 @@ namespace RacingLeagueHub.Application.Identity.Authentication.TwoFactorAuthentic
 
 public interface ITwoFactorService
 {
-    Task<TwoFactorSetupDto> StartSetupAsync(long userId, CancellationToken ct = default);
-    Task<ConfirmTwoFactorResponse> ConfirmSetupAsync(long userId, string code, CancellationToken ct = default);
-    Task<RecoveryCodesResponse> RegenerateRecoveryCodesAsync(long userId, CancellationToken ct = default);
+    Task<TwoFactorSetupDto> StartSetupAsync(int userId, CancellationToken ct = default);
+    Task<ConfirmTwoFactorResponse> ConfirmSetupAsync(int userId, string code, CancellationToken ct = default);
+    Task<RecoveryCodesResponse> RegenerateRecoveryCodesAsync(int userId, CancellationToken ct = default);
 }
