+++ RacingLeagueHub.Application/Teams/TeamService.cs
@@ -17,7 +17,7 @@ public class TeamService : ITeamService
         this.commands = commands;
     }
 
-    public async Task<TeamDto?> GetByIdAsync(long id, CancellationToken ct)
+    public async Task<TeamDto?> GetByIdAsync(int id, CancellationToken ct)
     {
         return await queries.GetByIdAsync(id, ct);
     }
@@ -32,12 +32,12 @@ public class TeamService : ITeamService
         return await commands.AddAsync(dto, ct);
     }
 
-    public async Task<TeamDto?> UpdateAsync(long id, UpdateTeamDto dto, CancellationToken ct)
+    public async Task<TeamDto?> UpdateAsync(int id, UpdateTeamDto dto, CancellationToken ct)
     {
         return await commands.UpdateAsync(id, dto, ct);
     }
 
-    public async Task<bool> DeleteAsync(long id, CancellationToken ct)
+    public async Task<bool> DeleteAsync(int id, CancellationToken ct)
     {
         return await commands.DeleteAsync(id, ct);
     }
