+++ RacingLeagueHub.Application/Tracks/ITrackService.cs
@@ -6,10 +6,10 @@ namespace RacingLeagueHub.Application.Tracks;
 
 public interface ITrackService
 {
-    Task<TrackDto?> GetByIdAsync(long id, CancellationToken ct);
+    Task<TrackDto?> GetByIdAsync(int id, CancellationToken ct);
     Task<PagedResult<TrackDto>> GetPagedAsync(int page, CancellationToken ct);
     Task<List<LookupDto>> GetLookupsAsync(CancellationToken ct);
     Task<TrackDto> AddAsync(CreateTrackDto dto, CancellationToken ct);
-    Task<TrackDto?> UpdateAsync(long id, UpdateTrackDto dto, CancellationToken ct);
-    Task<bool> DeleteAsync(long id, CancellationToken ct);
+    Task<TrackDto?> UpdateAsync(int id, UpdateTrackDto dto, CancellationToken ct);
+    Task<bool> DeleteAsync(int id, CancellationToken ct);
 }
