+++ RacingLeagueHub.Application/Tracks/Persistence/ITrackCommands.cs
@@ -5,6 +5,6 @@ namespace RacingLeagueHub.Application.Tracks.Persistence;
 public interface ITrackCommands
 {
     Task<TrackDto> AddAsync(CreateTrackDto dto, CancellationToken ct);
-    Task<TrackDto?> UpdateAsync(long id, UpdateTrackDto dto, CancellationToken ct);
-    Task<bool> DeleteAsync(long id, CancellationToken ct);
+    Task<TrackDto?> UpdateAsync(int id, UpdateTrackDto dto, CancellationToken ct);
+    Task<bool> DeleteAsync(int id, CancellationToken ct);
 }
