+++ RacingLeagueHub.Domain/Entities/GameTeam.cs
@@ -6,12 +6,12 @@ namespace RacingLeagueHub.Domain.Entities;
 public class GameTeam : EntityBase
 {
     public Game Game { get; set; }
-    public long TeamId { get; set; }
+    public int TeamId { get; set; }
     public string Name { get; set; }
     public string ShortName { get; set; }
     public string Abbreviation { get; set; }
     public string? Color { get; set; }
-    public long? LogoResourceId { get; set; }
+    public int? LogoResourceId { get; set; }
     public short TelemetryId { get; set; }
 
     public virtual Team Team { get; set; }
