+++ RacingLeagueHub.Domain/Entities/GrandsPrix/GrandPrix.cs
@@ -5,8 +5,8 @@ namespace RacingLeagueHub.Domain.Entities.GrandsPrix;
 
 public class GrandPrix : EntityBase
 {
-    public long TrackLayoutId { get; set; }
-    public long SeasonId { get; set; }
+    public int TrackLayoutId { get; set; }
+    public int SeasonId { get; set; }
     public string Name { get; set; }
     public DateTimeOffset StartingAt { get; set; }
     public string? VodUrl { get; set; }
