+++ RacingLeagueHub.Domain/Entities/UserExternalLogin.cs
@@ -2,7 +2,7 @@
 
 public class UserExternalLogin : EntityBase
 {
-    public long UserId { get; set; }
+    public int UserId { get; set; }
 
     public string Provider { get; set; } = null!;
     public string ProviderUserId { get; set; } = null!;
